Created HtmlToText class
