+++ lib/wgit.rb
@@ -5,6 +5,7 @@ require_relative 'wgit/logger'
 require_relative 'wgit/assertable'
 require_relative 'wgit/utils'
 require_relative 'wgit/url'
+require_relative 'wgit/html_to_text'
 require_relative 'wgit/document'
 require_relative 'wgit/document_extractors'
 require_relative 'wgit/crawler'
