+++ load.rb
@@ -8,6 +8,7 @@ load 'lib/wgit/logger.rb'
 load 'lib/wgit/assertable.rb'
 load 'lib/wgit/utils.rb'
 load 'lib/wgit/url.rb'
+load 'lib/wgit/html_to_text.rb'
 load 'lib/wgit/document.rb'
 load 'lib/wgit/document_extractors.rb'
 load 'lib/wgit/crawler.rb'
